fix : 페이지 이동시 탭바 옵션 초기값으로 수정
